+++ wutengine_examples/pong/src/spawn.rs
@@ -3,6 +3,7 @@
 use std::time::Instant;
 
 use wutengine::builtins::components::util::FramerateCounter;
+use wutengine::builtins::components::CameraType::{self};
 use wutengine::builtins::components::{Camera, Material, Mesh, Name, Transform};
 use wutengine::command::{Command, FullscreenType, OpenWindowParams};
 use wutengine::graphics::color::Color;
@@ -14,7 +15,7 @@ use wutengine::map;
 use wutengine::math::{vec3, Quat, Vec3};
 use wutengine::plugins::WutEnginePlugin;
 
-use crate::BallData;
+use crate::{BallData, CameraData};
 
 /// Plugin that only injects the initial components to get the game started
 pub(crate) struct PongStarterPlugin;
@@ -122,11 +123,15 @@ fn make_camera(commands: &mut Command) {
         .entity()
         .spawn()
         .with_component(Name::new("Camera"))
+        .with_component(CameraData {
+            start: Instant::now(),
+        })
         .with_component(FramerateCounter::new())
-        .with_component(Transform::with_pos(Vec3::new(0.0, 0.0, -1.0)))
+        .with_component(Transform::with_pos(Vec3::new(0.0, 0.0, -3.0)))
         .with_component(Camera {
             display: WindowIdentifier::new("main"),
             clear_color: Color::BLACK,
+            camera_type: CameraType::Perspective(75.0),
         })
         .build();
 }